move contextSourceFetchers to their own subdirectory
