+++ src/test/java/org/opensearch/neuralsearch/processor/factory/RerankProcessorFactoryTests.java
@@ -16,10 +16,10 @@ import org.junit.Before;
 import org.mockito.Mock;
 import org.opensearch.OpenSearchParseException;
 import org.opensearch.neuralsearch.ml.MLCommonsClientAccessor;
-import org.opensearch.neuralsearch.processor.rerank.DocumentContextSourceFetcher;
 import org.opensearch.neuralsearch.processor.rerank.MLOpenSearchRerankProcessor;
 import org.opensearch.neuralsearch.processor.rerank.RerankProcessor;
 import org.opensearch.neuralsearch.processor.rerank.RerankType;
+import org.opensearch.neuralsearch.processor.rerank.context.DocumentContextSourceFetcher;
 import org.opensearch.search.pipeline.Processor.PipelineContext;
 import org.opensearch.search.pipeline.SearchResponseProcessor;
 import org.opensearch.test.OpenSearchTestCase;
